+++ examples/shopping-list-tutorial/lesson-4/run.js
@@ -50,7 +50,7 @@ void (async () => {
         const resolveConfig = merge(defaultResolveConfig, appConfig, devConfig)
         await reset(resolveConfig, {
           dropEventStore: false,
-          dropEventBus: true,
+          dropEventSubscriber: true,
           dropReadModels: true,
           dropSagas: true,
         })