+++ functional-tests/app/common/aggregates/scheduler-test.commands.js
@@ -1,0 +1,10 @@
+export default {
+  requestSagaSchedulerTest: (_, { payload }) => ({
+    type: 'SagaSchedulerRequested',
+    payload,
+  }),
+  succeedSagaSchedulerTest: (_, { payload }) => ({
+    type: 'SagaSchedulerSucceeded',
+    payload,
+  }),
+}