+++ functional-tests/utils/constants.ts
@@ -1,0 +1,3 @@
+export const SAGA_TEST_EVENTS_COUNT = 100
+// eslint-disable-next-line spellcheck/spell-checker
+export const SAGA_TEST_AGGREGATE_ID = 'ubay72N4egmqqI95qPmjv'