+++ packages/core/core/src/saga/wrap-side-effects.ts
@@ -1,13 +1,14 @@
-import { SagaProperties, SideEffectsCollection } from './types'
+import { SideEffectsCollection } from '../types/core'
+import { SideEffectsContext } from './types'
 
 const sideEffect = async (
-  sagaProperties: SagaProperties,
   callback: Function,
   isEnabled: boolean,
+  sideEffectsContext: SideEffectsContext,
   ...args: any[]
 ) => {
   if (isEnabled) {
-    return await callback(...args, sagaProperties)
+    return await callback(...args, sideEffectsContext)
   } else {
     // Explicitly return undefined for disabled side-effects
     return undefined
@@ -15,9 +16,9 @@ const sideEffect = async (
 }
 
 export const wrapSideEffects = (
-  sagaProperties: SagaProperties,
   sideEffects: SideEffectsCollection,
-  isEnabled: boolean
+  isEnabled: boolean,
+  sideEffectsContext: SideEffectsContext
 ) => {
   return Object.keys(sideEffects).reduce<SideEffectsCollection>(
     (acc, effectName) => {
@@ -25,9 +26,9 @@ export const wrapSideEffects = (
       if (typeof effectOrSubCollection === 'function') {
         acc[effectName] = sideEffect.bind(
           null,
-          sagaProperties,
           effectOrSubCollection,
-          isEnabled
+          isEnabled,
+          sideEffectsContext
         )
       }
       if (
@@ -35,9 +36,9 @@ export const wrapSideEffects = (
         effectOrSubCollection !== null
       ) {
         acc[effectName] = wrapSideEffects(
-          sagaProperties,
           effectOrSubCollection,
-          isEnabled
+          isEnabled,
+          sideEffectsContext
         )
       }
       return acc