+++ packages/runtime/adapters/readmodel-adapters/readmodel-base/src/create-adapter.ts
@@ -32,10 +32,6 @@ const createAdapter = <
     subscribe,
     unsubscribe,
     resubscribe,
-    deleteProperty,
-    getProperty,
-    listProperties,
-    setProperty,
     resume,
     pause,
     reset,
@@ -53,7 +49,7 @@ const createAdapter = <
 
   if (Object.keys(restApi).length > 0) {
     throw new Error(
-      `Read-model adapter implementation should not provide extra methods: ${JSON.stringify(
+      `Read model adapter implementation should not provide extra methods: ${JSON.stringify(
         Object.keys(restApi)
       )}`
     )
@@ -73,10 +69,6 @@ const createAdapter = <
     subscribe,
     unsubscribe,
     resubscribe,
-    deleteProperty,
-    getProperty,
-    listProperties,
-    setProperty,
     resume,
     pause,
     reset,
@@ -88,7 +80,7 @@ const createAdapter = <
     StoreApi<AdapterPool>
   >) {
     if (typeof storeApi[key] !== 'function') {
-      throw new Error(`Store API method ${key} should be function`)
+      throw new Error(`Store API method ${key} should be a function`)
     }
   }
 
@@ -96,7 +88,7 @@ const createAdapter = <
     AdapterOperations<AdapterPool>
   >) {
     if (typeof adapterOperations[key] !== 'function') {
-      throw new Error(`Adapter operation method ${key} should be function`)
+      throw new Error(`Adapter operation method ${key} should be a function`)
     }
   }
 
@@ -116,10 +108,6 @@ const createAdapter = <
     subscribe: wrapOperation(pool, 'subscribe', subscribe),
     unsubscribe: wrapOperation(pool, 'unsubscribe', unsubscribe),
     resubscribe: wrapOperation(pool, 'resubscribe', resubscribe),
-    deleteProperty: wrapOperation(pool, 'deleteProperty', deleteProperty),
-    getProperty: wrapOperation(pool, 'getProperty', getProperty),
-    listProperties: wrapOperation(pool, 'listProperties', listProperties),
-    setProperty: wrapOperation(pool, 'setProperty', setProperty),
     resume: wrapOperation(pool, 'resume', resume),
     pause: wrapOperation(pool, 'pause', pause),
     reset: wrapOperation(pool, 'reset', reset),