+++ packages/runtime/adapters/readmodel-adapters/readmodel-postgresql-serverless/src/convert-result-row.ts
@@ -3,7 +3,7 @@ import type { ConvertResultRowMethod, RowLike } from './types'
 const convertResultRow: ConvertResultRowMethod = (inputRow, fieldList) => {
   if (fieldList != null && fieldList.constructor !== Object) {
     throw new Error(
-      'Field list should be object with enumerated selected fields'
+      'Field list should be an object with enumerated selected fields'
     )
   }
   const row: RowLike = { ...inputRow }