+++ packages/runtime/adapters/readmodel-adapters/readmodel-postgresql-serverless/src/index.ts
@@ -38,15 +38,12 @@ import inlineLedgerForceStop from './inline-ledger-force-stop'
 import generateGuid from './generate-guid'
 import isHighloadError from './is-highload-error'
 import isTimeoutError from './is-timeout-error'
+import maybeInit from './maybe-init'
 
 import dropReadModel from './drop-read-model'
 import subscribe from './subscribe'
 import resubscribe from './resubscribe'
 import unsubscribe from './unsubscribe'
-import deleteProperty from './delete-property'
-import getProperty from './get-property'
-import listProperties from './list-properties'
-import setProperty from './set-property'
 import reset from './reset'
 import pause from './pause'
 import resume from './resume'
@@ -80,6 +77,7 @@ const internalMethods: InternalMethods = {
   isTimeoutError,
   isHighloadError,
   dropReadModel,
+  maybeInit,
   escapeId,
   escapeStr,
 }
@@ -88,10 +86,6 @@ const externalMethods: ExternalMethods = {
   subscribe,
   resubscribe,
   unsubscribe,
-  deleteProperty,
-  getProperty,
-  listProperties,
-  setProperty,
   resume,
   pause,
   reset,