+++ packages/runtime/adapters/readmodel-adapters/readmodel-postgresql-serverless/src/subscribe.ts
@@ -19,6 +19,8 @@ const subscribe: ExternalMethods['subscribe'] = async (
   const ledgerTableNameAsId = escapeId(`__${schemaName}__LEDGER__`)
   try {
     pool.activePassthrough = true
+    await pool.maybeInit(pool)
+
     while (true) {
       try {
         await inlineLedgerForceStop(pool, readModelName)