+++ packages/runtime/runtime/test/cloud-entry.test.js
@@ -414,7 +414,7 @@ describe('Cloud entry', () => {
 
       expect(result.statusCode).toEqual(409)
       expect(result.headers).toEqual({ 'Content-Type': 'text/plain' })
-      expect(result.body).toContain('is not actual at the moment')
+      expect(result.body).toContain('is currently out of date')
     })
 
     test('should fail command via POST /"rootPath"/api/commands/ with CommandError', async () => {